requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
crispy-bootstrap5==2024.10
    # via -r requirements.in
django==5.1.2
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-crispy-forms
django-crispy-forms==2.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
sqlparse==0.5.1
    # via django
